[chore] new tests: ordered delivery (#15)
Add a test that does out-of-order delivery of events. This is attempting
to reproduce a problem.

Also: move the definition of `Event` (unchanged)
Also: add some documentation to existing functions
Also: move the notification registrations for tests inside the
corresponding tests so that if they aren't being run, there might not
need to be a broadcast consumer
Also: adjust some constants to try to make tests more reliable