[chore] refactor broadcast consumer group allocation (#28)
This is the first change in a set of changes that are needed to allow
use of the events system without a database. In this PR, the way that
broadcast consumer groups are allocated changes such that we
double-check that we have exclusive use of the consumer group.

This is incomplete because it still depends upon holding a database
lock. Without a database lock, there is a race during the time a reader
is restarted.

A followup change will fix that race.

Specific changes in this PR:

- broadcast consumer group numbers are allocated randomly instead of
sequentially with a bias towards small numbers
- exclusive use of a broadcast consumer group is robustly verified with
Kafka
- the set of errors that are treated as transient when getting a
consumer group is expanded, thus increasing reliability with respect to
Kafka being in the middle of something during program startup
- A few bits of code were simplified.
- The batch delivery test was split up into three tests (one for each
consumer type)

Part 2: #31 
Part 3: #30