[feat] refresh broadcast readers on reset (#31)
This is part2 of adding support for using events w/o a database. Here we
further refactor broadcast consumers so that when they're re-created,
they are still guaranteed to be exclusive.

One side effect of this work is that topics are created with
`message.timestamp.type`: `LogAppendTime`

This is on top of #28, followed by #30