[feat] Add easy-to-use wrapper for creating event library instances without a database (#30)
Now can use `eventnodb` if you want events without a database. Without a
database, `ConsumeExactlyOnce` is not supported.

Additional change: when produce gets an UnknownTopicOrPartition error,
it now tries again, trying each broker in turn. This additional change
was required to make the tests more reliable.

Additional change: starting up Kafka for testing is now documented.

This is on top of #28 and #31