[fix] when consumer group allocation fails, retry the same group except for unavailable errors (#67)
Change the consumer group allocation loop: if the error is that the
group is in use, try a different group. If the error is anything else,
retry with the same group.

This assumes that the errors are correctly classified. If the group is
in use but we return the wrong error, this will cause a loop.